WIP, separate out filters into their own component
